GCI-142 Remove redundant value unboxing (#2398)

+++ api/src/main/java/org/openmrs/api/db/hibernate/ChainingInterceptor.java
@@ -178,7 +178,7 @@ public class ChainingInterceptor implements Interceptor {
 		int[] uniquePrimitiveIndices = new int[uniqueIndices.size()];
 		
 		for (int x = 0; x < uniqueIndices.size(); x++) {
-			uniquePrimitiveIndices[x] = uniqueIndices.get(x).intValue();
+			uniquePrimitiveIndices[x] = uniqueIndices.get(x);
 		}
 		
 		return uniquePrimitiveIndices;
