+++ api/src/main/java/org/openmrs/util/DoubleRange.java
@@ -144,18 +144,18 @@ public class DoubleRange implements Comparable<DoubleRange> {
 	@Override
 	public String toString() {
 		StringBuilder ret = new StringBuilder();
-		if (low != null && low.doubleValue() != Double.NEGATIVE_INFINITY) {
+		if (low != null && low != Double.NEGATIVE_INFINITY) {
 			ret.append(">");
 			if (closedLow) {
 				ret.append("=");
 			}
 			ret.append(" ").append(Format.format(low));
-			if (high != null && high.doubleValue() != Double.NEGATIVE_INFINITY) {
+			if (high != null && high != Double.NEGATIVE_INFINITY) {
 				//BUG: should not append this if high is also infinite
 				ret.append(" and ");
 			}
 		}
-		if (high != null && high.doubleValue() != Double.POSITIVE_INFINITY) {
+		if (high != null && high != Double.POSITIVE_INFINITY) {
 			ret.append("<");
 			if (closedHigh) {
 				ret.append("=");
